Widen VB.NET set element types for mixed-width ints
VB rejects integer literals outside the Integer range with
``BC30439: Constant expression not representable in type 'Integer'``
when the surrounding collection is typed ``HashSet(Of Integer)``.
Map ``WideInt`` to ``Long`` in the VB typed-opener config and add a
``format_integer_widened`` that always appends the ``L`` suffix so
``HashSet(Of Long) From { 1L, 1099511627776L }`` compiles.

+++ tests/integration/cases/set_mixed_int_widths/VisualBasic_combined.vb
@@ -1,15 +1,15 @@
 Module Check
     Sub _declaration()
-        Dim my_data = New HashSet(Of Integer) From {
-            1,
-            1099511627776
+        Dim my_data = New HashSet(Of Long) From {
+            1L,
+            1099511627776L
         }
     End Sub
     Sub _assignment()
         Dim my_data As Object
-        my_data = New HashSet(Of Integer) From {
-            1,
-            1099511627776
+        my_data = New HashSet(Of Long) From {
+            1L,
+            1099511627776L
         }
     End Sub
 End Module
